CSF: Refactor to stabilize export to storyName transform
