+++ lib/client-api/src/client_api.ts
@@ -1,7 +1,6 @@
 /* eslint no-underscore-dangle: 0 */
 import deprecate from 'util-deprecate';
 import isPlainObject from 'is-plain-object';
-import startCase from 'lodash/startCase';
 import { logger } from '@storybook/client-logger';
 import addons, { StoryContext, StoryFn, Parameters, OptionsParameter } from '@storybook/addons';
 import Events from '@storybook/core-events';
@@ -83,8 +82,6 @@ const withSubscriptionTracking = (storyFn: StoryFn) => {
   return result;
 };
 
-export const makeDisplayName = (key: string) => startCase(key);
-
 export default class ClientApi {
   private _storyStore: StoryStore;
 
@@ -125,8 +122,6 @@ export default class ClientApi {
       },
       this._globalParameters.options
     );
-
-  getMakeDisplayName = () => makeDisplayName;
 
   addDecorator = (decorator: DecoratorFunction) => {
     this._globalDecorators.push(decorator);
