+++ lib/core/src/client/preview/start.js
@@ -7,7 +7,7 @@ import AnsiToHtml from 'ansi-to-html';
 import addons from '@storybook/addons';
 import createChannel from '@storybook/channel-postmessage';
 import { ClientApi, StoryStore, ConfigApi } from '@storybook/client-api';
-import { toId } from '@storybook/router/utils';
+import { toId, storyNameFromExport } from '@storybook/router/utils';
 import { logger } from '@storybook/client-logger';
 import Events from '@storybook/core-events';
 
@@ -398,9 +398,6 @@ export default function start(render, { decorateStory } = {}) {
       // We pass true here to avoid the warning about HMR. It's cool clientApi, we got this
       const kind = clientApi.storiesOf(kindName, true);
 
-      // Transform the CSF named export if the user hasn't specified a name
-      const makeDisplayName = clientApi.getMakeDisplayName();
-
       // we should always have a framework, rest optional
       kind.addParameters({ framework, component, ...params });
 
@@ -418,7 +415,7 @@ export default function start(render, { decorateStory } = {}) {
           }
           const decoratorParams = decorators ? { decorators } : null;
           const displayNameParams = name ? { displayName: name } : {};
-          kind.add(makeDisplayName(key), storyFn, {
+          kind.add(storyNameFromExport(key), storyFn, {
             ...parameters,
             ...decoratorParams,
             ...displayNameParams,
