+++ lib/router/src/utils.test.js
@@ -1,4 +1,4 @@
-import { toId } from '../utils';
+import { toId, storyNameFromExport } from './utils';
 
 describe('toId', () => {
   [
@@ -35,5 +35,19 @@ describe('toId', () => {
 
   it('does not allow empty story', () => {
     expect(() => toId('kind', '')).toThrow(`Invalid name '', must include alphanumeric characters`);
+  });
+});
+
+describe('storyNameFromExport', () => {
+  it('should format CSF exports with sensible defaults', () => {
+    const testCases = {
+      name: 'Name',
+      someName: 'Some Name',
+      someNAME: 'Some NAME',
+      some_custom_NAME: 'Some Custom NAME',
+      someName1234: 'Some Name 1234',
+      someName1_2_3_4: 'Some Name 1 2 3 4',
+    };
+    Object.entries(testCases).forEach(([key, val]) => expect(storyNameFromExport(key)).toBe(val));
   });
 });
